[style][backend] Move SupplierRequestDTO do diretório dto, para o diretório dto/requestDTO
